@markacola 2021-10-05
I also noted that there is no `using` around the `IDisposable` coming back from `Read`; Did you want me to add that in too? i.e.
```
using var file = reader.Read(stream);
if (!reader.IsMatch(file))
{
    candidates.Remove((FileFormat)reader);
}
```